Quick fix: I broke FuncLib
Good thing there aren't any users this time of night. In attempting to
make the functions prettier I did an "End if" instead of an "End
Function", and I forgot to "Call" one of the new functions in a
depreciated function. This is why we should test before we deploy.