NOTES - INTAKE DOCS RECEIVED.vbs
I think I added in the part about needing to be in a PRISM case, but you need you to look over.  I also added the date in.  The only thing that might be missing, is if there is no case in my globals, and I manually enter in a valid prism case number into my dialog box, it does not navigate to that case and write the note??  Am I still missing something??  Thanks!