NOTES - CSENET Info.vbs
This is a script that will copy info from INTD and paste it into a new CAAD note.  The only time we are running into issues, is when the text uses all of the lines in INTD, it will not paste nicely into CAAD lines.  Need to be able to word wrap, or just not use the script when the INTD screen has verbiage in paragraph form.  Thanks!