NOTES - ROP DETAIL.vbs
This is a CAAD note script that will note the dates and place a ROP was signed.  It will also message box if the CP or the NCP was under 18 yrs old when the ROP was signed.  It will CAAD note if one of the parties was under 18 yrs when the ROP was signed.