NOTES - COURT ORDER REQUEST.vbs
This is a CAAD note script that notes date court order was requested, place/person requested from, request method, court file number, order type, and worker signature.  It will then navigate to CAWT and create a worklist to remind you if the court order was received.  It does not save the worklist, in case you want to add additional notes in the worklist first before saving and also a date must be added.

I originally pull requested this at the training in March, but have since added the auto-filling of the PRISM case number.  I think this should be good to go!