NOTES - WAIVER OF PERSONAL SERVICE.vbs
This is a CAAD note script that will CAAD note when the waiver of personal service is signed by the CP.  It uses CAAD code D5010 and by using that CAAD code a worklist is automatically created.

This is another script I had originally pull requested at training in March, but have since added the logic to populate the PRISM case number when in a PRISM case.  I think it is done and working great!