Create BULK - EVALUATE NOCS.vbs
This script performs the following actions...

Navigates to USWT to grab the worker ID and name. It presents that information in a dialog, allowing the user to use this script on any authorized caseload. Once the worker has determined the caseload to work on, the script builds a multi-dimensional array of all cases with a D0800 worklist item. It then navigates to CALI to determine the program code. It then navigates to CAST to determine if the case is full service. It then navigates to CATH to determine the program code change date. It then navigates to GCSC to determine if the case is coop or non-coop.

The script then dumps all the information into an Excel file to present to the worker. For each case in the array, the script presents a dialog with the information, allowing the worker to determine which DORD doc to send, whether or not to create a FREE worklist item (along with the date and contents of the worklist item), the contents of a FREE CAAD note, and whether or not to purge.

To get to the next case, the worker must either SKIP the case, or sign the authorization field. There are additional mandatory conditions (the worker cannot select 2 DORD docs to send, the worker must include a date if they are creating a FREE worklist item). As the worker jumps to the next case, the Excel file is updated in real time with any changes. The row for the active case is highlighted yellow. When the worker jumps to the next case, the previous row is returned to a white back fill.

When the worker has reviewed all the cases, a dialog is presented giving the worker the ability to review the cases one more time or to continue. If the worker selects RETRY, the dialogs for each case are presented (do/loop). If the worker selects OK, the script goes to send the appropriate DORD docs, to create FREE worklists (as needed), and to create FREE CAAD notes (as needed), and purges the D0800 worklist items as selected.

At the end, the script_end_procedure includes a number of the purged D0800 worklist items. This allows agencies keeping stats to keep track of the number of purged worklist items and the time for the purge. (Quality control, yo.)