Create new file
This script allows the user to search for a CP or NCP on a particular CALI list.    The user is able to search by first name, last name, or both.  There is no soundex for the search.  Results do include partial matches.  If the user indicates a match is found, the script navigates to the CAST screen for the case.  The user can elect to run the search from the CALI associated with their log on or specify an office, unit, and position for the search.