E4111 Worklist Scrubber
This script navigates to USWT and filters for the E4111 Worklist, which notifies the case worker that the driver's license payment plan on the case is continuingly delinquent.  The script selects a worklist to navigate to PAPD, and then navigates to ENFL and filters for DLS records.  The script then searches the records for the case number and reads the DL status for the case.  If the status is ACT, that is, the NCP's driver's license is already suspended, the script returns to the USWT and purges the worklist.  If the DL status is anything else, the script does not purge the worklist.  The script then moves to the next E4111 worklist until all E4111 worklists have been evaluated.