IW caad cawt
This is for when you are sending an new or amended IW notice.  It creates a CAAD note that includes employer, new or amended IW, and includes, monthly, biweekly, semimonthly, and weekly amounts.  It gives a option for a CAWT note to be added 30 days in the future to check that iw started and adds the monthly, biweekly, semimonthly, and weekly amounts.