Create BULK - EMANCIPATION WORKLIST SCRUBBER.vbs
I updated this script as requested. This script looks for work list M0935 on a specific caseload. The script verifies there is a child that will be turning 18 years old within 3 months on each case. It verifies there is EMC language on suod. If there is no EMC language then a note is added to the M0935 work list that further review is needed. If EMC language is found then script sends dord doc f0300 and f0302 and purges the work list. Minimal testing has been done.