Resolves #355
This script can assist with tasks involved when a case is sanctioned for non-cooperation with child support and can assist with tasks involved when the sanction is cured.  The script can update GCSC screen, set file location on CAST, create a Word memo doc (Anoka-only), enter CAAD note, create/delete reminder worklists, and send a DORD-based memo to the custodial parent about the sanction.