Update ACTIONS - ESTB DORD DOCS FOR PA CASE.vbs
This is an updated script that will print the DORD docs for both parties when starting a new establish action on a PA case. It prints the Financial Statements, Waivers, Important Statement of Right & the NCP Notice of Liability. It writes a CAAD note of sending forms to parties and the CAWT note to track the return of the forms.