Resolved #404: AGENCY CUSTOM is live
This one was originally created by the great @lpshanley from St. Louis
County. It's been modified a bit since he originally dreamed it up, but
it's largely his invention.

This script allows a scriptwriter to "drop" a script into a folder and
make it available for local use. This is great for testing. The Power
Pad has been updated to allow for this (although agencies will have to
update their scripts to gain access, and that tool is coming shortly,
gang).