Resolves #473
This record income withholding information script allows the user to enter a CAAD note, create a worklist, or display in a message box the withholding amount broken down by pay frequency.  Anoka County uses this script when sending out an IW notice, and the CAAD note and worklist options are automatically selected.  The worklist is set for follow up after 30 days.  If the user de-selects the CAAD note and worklist options, the output will display in the message box.