Update unreimbursed-uninsured-returned-docs.vbs
previous pull request #593 only had partial code.  full code now added to the script to fix this error. 
New script deals with process when a party sends in a request to collect unreimbursed and uninsured expenses. This script creates a caad note of documents returned, sends 2 dord documents to the non requesting party, and create jude, obbd, ncod or cpod.