Update hearing-notes.vbs
Made slight adjustments to the list. @KallistaImdieke, note that it includes the original "initial version" as well as your update. This is done so that it's easy for the changelog function to see how many "new" things to display. In other words: always add your change to the top of the list and keep everything else that used to be there. :)