Updated Request for Enhancements #759
I updated the dialog box to have the option to enter worker's name, county, and phone number. When this is entered into CAAD, I have updated the script to write this info on one line.
I changed the word "Reviewing County" to "Responding County" on the dialog box.
I lastly added the option to have the user be able to add a CAWT note to follow up with maintaining county request.